+++ components/Footer.tsx
@@ -1,55 +1,31 @@
-import { useEffect } from "react"
-import { useFirestore } from "@/hooks/useFirestore"
+import React, { useEffect } from "react"
+import { useLogin } from "@/hooks/useLogin"
 import { signIn, signOut } from "next-auth/react"
-import CapturePhotoButton from "./CapturedPhotoButton"
-import { WeatherInfoDto } from "@/types/WeatherInfoDto"
 
-type Props = {
-    weather: WeatherInfoDto | undefined
-    prefecture: string
-}
-
-export default function Footer({ weather, prefecture }: Props) {
-    const { session, userCollection, handleUserCollection } = useFirestore()
+export default function Footer() {
+    const { session, userCollection, handleUserCollection } = useLogin()
     useEffect(() => {
         handleUserCollection()
     }, [session])
 
     return (
-        <footer
-            className={"fixed flex justify-center w-full h-[10vh] bottom-0"}
-        >
-            <div className="flex justify-between items-center w-[94%] h-[75%] rounded-2xl bg-gray-300 px-6">
-                <img src={"/images/home.svg"} className={"h-[70%]"} />
-                <CapturePhotoButton weather={weather} prefecture={prefecture} />
-                <img src={"/images/record.svg"} className={"h-[70%]"} />
-                {userCollection ? (
-                    <>
-                        <button
-                            className={"bg-white"}
-                            onClick={() => signOut()}
-                        >
-                            Sign outボタン
-                        </button>
-                        <button className={"h-[70%]"}>
-                            <img
-                                src={userCollection.image}
-                                alt={userCollection.name}
-                                className={"h-full rounded-full"}
-                            />
-                        </button>
-                    </>
-                ) : (
-                    <>
-                        <button className={"h-full"} onClick={() => signIn()}>
-                            <img
-                                src={"/images/user.svg"}
-                                className={"h-[70%]"}
-                            />
-                        </button>
-                    </>
-                )}
-            </div>
+        <footer>
+            {/* <div className="flex w-[94%] h-[75%] bg-black m-auto mt-[3%]"></div> */}
+            {userCollection ? (
+                <>
+                    <button onClick={() => signOut()}>Sign outボタン</button>
+                    <br />
+                    Signed in as {userCollection.email} <br />
+                    <p>name: {userCollection.name}</p>
+                    image:
+                    <img src={userCollection.image} alt={userCollection.name} />
+                </>
+            ) : (
+                <>
+                    Not signed in <br />
+                    <button onClick={() => signIn()}>Sign inボタン</button>
+                </>
+            )}
         </footer>
     )
 }
