--- components/atoms/ExpressionMark.tsx
@@ -1,7 +1,0 @@
-export default function ExpressionMark(emoji: string) {
-    return (
-        <li className={"flex w-1/6 items-center justify-center text-2xl"}>
-            {emoji}
-        </li>
-    )
-}
