+++ hooks/useLogin.tsx
@@ -1,10 +1,9 @@
-import React, { useState } from "react"
-import { useSession } from "next-auth/react"
-import { UserFirestoreDto } from "@/types/UserFirestoreDto"
+import React, { useEffect, useRef, useState } from "react"
+import { signIn, signOut, useSession } from "next-auth/react"
+import { UserFirestoreDto } from "@/types/userFirestoreDto"
 import axios from "axios"
-import { ExpWeaFirestoreDto } from "@/types/ExpWeaFirestoreDto"
 
-export const useFirestore = () => {
+export const useLogin = () => {
     const session = useSession()
     const [userCollection, setUserCollection] = useState<UserFirestoreDto>()
 
@@ -31,19 +30,5 @@ export const useFirestore = () => {
         })
     }
 
-    const insertCapturedPhoto = async (ex: ExpWeaFirestoreDto) => {
-        await axios.post("/api/expWea", {
-            faceImage: ex.faceImage,
-            email: ex.email,
-            expressions: ex.expressions,
-            weather: ex.weather,
-        })
-    }
-
-    return {
-        session,
-        userCollection,
-        handleUserCollection,
-        insertCapturedPhoto,
-    } as const
+    return { session, userCollection, handleUserCollection } as const
 }
