--- types/ExpWeaFirestoreDto.ts
@@ -1,14 +1,0 @@
-import { ExpressionsDto } from "./ExpressionsDto"
-
-export type ExpWeaFirestoreDto = {
-    faceImage: string
-    email: string
-    expressions: ExpressionsDto
-    weather: {
-        location: string
-        weather: string
-        temp: number
-        humidity: number
-        pressure: number
-    }
-}
