+++ types/expWheFirestoreDto.ts
@@ -1,0 +1,19 @@
+export type ExpWheFirestoreDto = {
+    faceImage: string
+    expressions: {
+        angry: number
+        disgusted: number
+        fearful: number
+        happy: number
+        neutral: number
+        sad: number
+        surprised: number
+    }
+    whether: {
+        location: string
+        whether: string
+        temp: number
+        humidity: number
+        pressure: number
+    }
+}
